requirements.txt
openai>=1.0.0
streamlit>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10